Improved granularity of event system
Looping flag and removal queue are now stored per event rather than being shared by all events.
This fixes an edge case where a removed event handler is still called because it was removed during a mouse click or similar.
It also allows us to detect when a handler fires its own event (which is an error and would lead to a stack overflow or infinite loop) while still allowing one event to be fired from within another.